added transitions v2
added transitions for missing data and
edited the way d3 uses the data to ensure
a smooth transition.

also added a selection in the drop down menu
for no missing data visualization

added a second drop down menu for changing
into different view. the first one is for
making all data visible. the other ones are not
yet implemented.